+++ types/Tax/Registrations.d.ts
@@ -889,7 +889,7 @@ declare module 'stripe' {
             /**
              * Type of registration in the US.
              */
-            type: 'state_sales_tax';
+            type: Us.Type;
           }
 
           namespace Us {
@@ -906,6 +906,8 @@ declare module 'stripe' {
                */
               jurisdiction: string;
             }
+
+            type Type = 'local_lease_tax' | 'state_sales_tax';
           }
 
           interface Za {