+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Changelog
 
+## 12.10.0 - 2023-06-22
+* [#1820](https://github.com/stripe/stripe-node/pull/1820) Update generated code
+  * Add support for `on_behalf_of` on `Mandate`
+* [#1817](https://github.com/stripe/stripe-node/pull/1817) Update README.md
+* [#1819](https://github.com/stripe/stripe-node/pull/1819) Update generated code
+  * Release specs are identical.
+* [#1813](https://github.com/stripe/stripe-node/pull/1813) Update generated code
+  * Change type of `Checkout.Session.success_url` from `string` to `string | null`
+  * Change type of `FileCreateParams.file` from `string` to `file`
+* [#1815](https://github.com/stripe/stripe-node/pull/1815) Generate FileCreateParams
+
 ## 12.10.0-beta.2 - 2023-06-15
 * [#1818](https://github.com/stripe/stripe-node/pull/1818) Update generated code for beta
   * Add support for `payment_details` on `ChargeCaptureParams`, `ChargeUpdateParams`, `PaymentIntentCaptureParams`, `PaymentIntentConfirmParams`, `PaymentIntentCreateParams`, `PaymentIntentUpdateParams`, and `PaymentIntent`
   * Add support for `statement_details` on `PaymentIntent.payment_method_options.card`, `PaymentIntentConfirmParams.payment_method_options.card`, `PaymentIntentCreateParams.payment_method_options.card`, and `PaymentIntentUpdateParams.payment_method_options.card`
 * [#1814](https://github.com/stripe/stripe-node/pull/1814) Update generated code for beta
-
 
 ## 12.10.0-beta.1 - 2023-06-08
 * [#1811](https://github.com/stripe/stripe-node/pull/1811) Update generated code for beta